Refactor code to use separate package for commands
